--- frontend/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    rules: {
-        '@typescript-eslint/no-unused-vars': 'off',
-        '@typescript-eslint/no-explicit-any': 'off',
-    }
-};
